requirements.txt
aiodns>=3.0.0
aiohttp>=3.8.0
uvloop>=0.17.0; sys_platform != 'win32'
rich>=13.0.0
pyfiglet>=0.8.0
openai>=1.0.0
python-dotenv>=1.0.0
readchar>=4.0.0
jinja2
dnspython>=2.6.0
